requirements.txt
appnope==0.1.3
asttokens==2.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
executing==2.0.1
flake8==7.1.0
icalendar==5.0.13
idna==3.4
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.6
mccabe==0.7.0
parso==0.8.3
pexpect==4.8.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.32.3
six==1.16.0
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.12.2
urllib3==2.1.0
wcwidth==0.2.10
